third_party/blink/web_tests/virtual/select-parser-relaxation-opt-out/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-click-drag-option.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Clicking the invoker button and dragging to an option should choose that option and close the picker.
  assert_equals: appearance:base-select must be supported in order to run this test. expected "base-select" but got "auto"
[FAIL] Releasing the pointer on an option positioned on top of the invoker button should not close the picker.
  assert_equals: appearance:base-select must be supported in order to run this test. expected "base-select" but got "auto"
Harness: the test ran to completion.
